requirements.txt
streamlit==1.45.1
tensorflow-cpu==2.19.0
numpy==1.26.4
scipy==1.15.3
Pillow==11.2.1
matplotlib==3.10.3
